Move assumptions related tests to separate modules
