--- assertj-core/src/it/setup/invoker.properties
@@ -1,1 +1,0 @@
-invoker.goals = install
