+++ assertj-tests/assertj-integration-tests/assertj-core-junit4-with-opentest4j/pom.xml
@@ -3,27 +3,30 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>it</groupId>
-    <artifactId>setup</artifactId>
-    <version>0</version>
+    <groupId>org.assertj</groupId>
+    <artifactId>assertj-integration-tests</artifactId>
+    <version>3.24.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>testng-with-junit4</artifactId>
+  <artifactId>assertj-core-junit4-with-opentest4j</artifactId>
+
+  <name>AssertJ Core - JUnit 4 with opentest4j</name>
 
   <dependencies>
-    <dependency>
-      <groupId>@project.groupId@</groupId>
-      <artifactId>@project.artifactId@</artifactId>
-      <scope>test</scope>
-    </dependency>
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>org.testng</groupId>
-      <artifactId>testng</artifactId>
+      <groupId>org.assertj</groupId>
+      <artifactId>assertj-core</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.opentest4j</groupId>
+      <artifactId>opentest4j</artifactId>
       <scope>test</scope>
     </dependency>
   </dependencies>
