+++ assertj-tests/assertj-integration-tests/assertj-core-junit4-with-opentest4j/src/test/java/org/assertj/core/tests/JUnit4_with_opentest4j_Test.java
@@ -10,16 +10,16 @@
  *
  * Copyright 2012-2022 the original author or authors.
  */
-package maven.invoker.it.assumptions;
+package org.assertj.core.tests;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.assertj.core.api.AssumptionExceptionFactory.getPreferredAssumptionException;
 import static org.assertj.core.api.Assumptions.assumeThat;
+import static org.assertj.core.api.Assumptions.setPreferredAssumptionException;
 import static org.assertj.core.api.BDDAssertions.then;
 
-import org.assertj.core.api.Assumptions;
 import org.assertj.core.configuration.PreferredAssumptionException;
 import org.junit.After;
 import org.junit.AssumptionViolatedException;
@@ -33,15 +33,7 @@ public class JUnit4_with_opentest4j_Test {
   @After
   public void afterEachTest() {
     // reset to the default value to avoid side effects on the other tests
-    Assumptions.setPreferredAssumptionException(DEFAULT_PREFERRED_ASSUMPTION_EXCEPTION);
-  }
-
-  @Test
-  public void should_throw_JUnit4_AssumptionViolatedException_when_assumption_fails() {
-    // WHEN
-    Throwable thrown = catchThrowable(() -> assumeThat(true).isFalse());
-    // THEN
-    then(thrown).isInstanceOf(AssumptionViolatedException.class);
+    setPreferredAssumptionException(DEFAULT_PREFERRED_ASSUMPTION_EXCEPTION);
   }
 
   @Test
@@ -63,14 +55,21 @@ public class JUnit4_with_opentest4j_Test {
   }
 
   @Test
+  public void should_throw_AssumptionViolatedException_when_assumption_fails() {
+    // WHEN
+    Throwable thrown = catchThrowable(() -> assumeThat(true).isFalse());
+    // THEN
+    then(thrown).isInstanceOf(AssumptionViolatedException.class);
+  }
+
+  @Test
   public void should_throw_TestAbortedException_when_assumption_fails_if_preferredAssumptionException_is_set_to_opentest4j() {
     // GIVEN
-    Assumptions.setPreferredAssumptionException(PreferredAssumptionException.JUNIT5);
+    setPreferredAssumptionException(PreferredAssumptionException.JUNIT5);
     // WHEN
     Throwable thrown = catchThrowable(() -> assumeThat(true).isFalse());
     // THEN
     then(thrown).isInstanceOf(TestAbortedException.class);
   }
-  
-  
+
 }
