+++ assertj-tests/assertj-integration-tests/assertj-core-testng-with-junit4/pom.xml
@@ -3,27 +3,30 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>it</groupId>
-    <artifactId>setup</artifactId>
-    <version>0</version>
+    <groupId>org.assertj</groupId>
+    <artifactId>assertj-integration-tests</artifactId>
+    <version>3.24.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>junit4-with-opentest4j</artifactId>
+  <artifactId>assertj-core-testng-with-junit4</artifactId>
+
+  <name>AssertJ Core - TestNG with JUnit 4</name>
 
   <dependencies>
-    <dependency>
-      <groupId>@project.groupId@</groupId>
-      <artifactId>@project.artifactId@</artifactId>
-      <scope>test</scope>
-    </dependency>
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>org.opentest4j</groupId>
-      <artifactId>opentest4j</artifactId>
+      <groupId>org.assertj</groupId>
+      <artifactId>assertj-core</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.testng</groupId>
+      <artifactId>testng</artifactId>
       <scope>test</scope>
     </dependency>
   </dependencies>
