+++ assertj-tests/assertj-integration-tests/assertj-core-testng-with-junit4/src/test/java/org/assertj/core/tests/TestNG_with_JUnit4_Test.java
@@ -10,39 +10,30 @@
  *
  * Copyright 2012-2022 the original author or authors.
  */
-package maven.invoker.it.assumptions;
+package org.assertj.core.tests;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.assertj.core.api.AssumptionExceptionFactory.getPreferredAssumptionException;
 import static org.assertj.core.api.Assumptions.assumeThat;
+import static org.assertj.core.api.Assumptions.setPreferredAssumptionException;
 import static org.assertj.core.api.BDDAssertions.then;
 
-import org.assertj.core.api.Assumptions;
-import org.assertj.core.configuration.Configuration;
 import org.assertj.core.configuration.PreferredAssumptionException;
 import org.junit.AssumptionViolatedException;
 import org.testng.SkipException;
-import org.testng.annotations.AfterMethod;  
+import org.testng.annotations.AfterMethod;
 import org.testng.annotations.Test;
 
 public class TestNG_with_JUnit4_Test {
-  
+
   private static final PreferredAssumptionException DEFAULT_PREFERRED_ASSUMPTION_EXCEPTION = getPreferredAssumptionException();
 
   @AfterMethod
   public void afterEachTest() {
     // reset to the default value to avoid side effects on the other tests
-    Assumptions.setPreferredAssumptionException(DEFAULT_PREFERRED_ASSUMPTION_EXCEPTION);
-  }
-
-  @Test
-  public void should_throw_TestNG_SkipException_when_assumption_fails() {
-    // WHEN
-    Throwable thrown = catchThrowable(() -> assumeThat(true).isFalse());
-    // THEN
-    then(thrown).isInstanceOf(SkipException.class);
+    setPreferredAssumptionException(DEFAULT_PREFERRED_ASSUMPTION_EXCEPTION);
   }
 
   @Test
@@ -58,19 +49,27 @@ public class TestNG_with_JUnit4_Test {
   }
 
   @Test
+  public void should_not_have_opentest4j_in_the_classpath() {
+    // WHEN/THEN
+    assertThatExceptionOfType(ClassNotFoundException.class).isThrownBy(() -> Class.forName("org.opentest4j.TestAbortedException"));
+  }
+
+  @Test
+  public void should_throw_SkipException_when_assumption_fails() {
+    // WHEN
+    Throwable thrown = catchThrowable(() -> assumeThat(true).isFalse());
+    // THEN
+    then(thrown).isInstanceOf(SkipException.class);
+  }
+
+  @Test
   public void should_throw_AssumptionViolatedException_when_assumption_fails_if_preferredAssumptionException_is_set_to_JUnit4() {
     // GIVEN
-    Assumptions.setPreferredAssumptionException(PreferredAssumptionException.JUNIT4);
+    setPreferredAssumptionException(PreferredAssumptionException.JUNIT4);
     // WHEN
     Throwable thrown = catchThrowable(() -> assumeThat(true).isFalse());
     // THEN
     then(thrown).isInstanceOf(AssumptionViolatedException.class);
-  }
-  
-  @Test
-  public void should_not_have_opentest4j_in_the_classpath() {
-    // WHEN/THEN
-    assertThatExceptionOfType(ClassNotFoundException.class).isThrownBy(() -> Class.forName("org.opentest4j.TestAbortedException"));
   }
 
 }
