+++ assertj-tests/assertj-integration-tests/pom.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.assertj</groupId>
+    <artifactId>assertj-tests</artifactId>
+    <version>3.24.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>assertj-integration-tests</artifactId>
+  <packaging>pom</packaging>
+
+  <name>AssertJ Integration Tests</name>
+
+  <modules>
+    <module>assertj-core-junit4-with-opentest4j</module>
+    <module>assertj-core-testng-with-junit4</module>
+  </modules>
+
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.testng</groupId>
+        <artifactId>testng</artifactId>
+        <version>7.6.1</version>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+
+</project>
