+++ assertj-tests/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.assertj</groupId>
+    <artifactId>assertj-parent</artifactId>
+    <version>3.24.0-SNAPSHOT</version>
+    <relativePath>../assertj-parent</relativePath>
+  </parent>
+
+  <artifactId>assertj-tests</artifactId>
+  <packaging>pom</packaging>
+
+  <name>AssertJ Tests</name>
+
+  <modules>
+    <module>assertj-integration-tests</module>
+  </modules>
+
+</project>
