+++ pom.xml
@@ -92,6 +92,7 @@
   <modules>
     <module>assertj-parent</module>
     <module>assertj-core</module>
+    <module>assertj-tests</module>
   </modules>
 
   <scm>
